Create PaymentSummary by splitting up PaymentPicker.
Show message if no active payment options.
